Move `currentIncidents` to `incidents.go` file
